requirements.txt
pydantic>=2.0
pyyaml>=6.0
pyyaml
pandas
openpyxl
pyarrow>=14.0
torch
matplotlib